fix warnings
git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@5386 a380766d-ff14-0410-b294-a243070f3f08

+++ source/lang/LangSource/GC.cpp
@@ -820,7 +820,7 @@ bool PyrGC::ListSanity()
 	bool found;
 	
 	if (StackDepth() < 0) {
-		fprintf(stderr, "stack underflow %ld\n", StackDepth());
+		fprintf(stderr, "stack underflow %d\n", (int)StackDepth());
 		return false;
 	}
